+++ src/editor-service/implementations/index.ts
@@ -1,0 +1,2 @@
+export * from "./EditorService2";
+export * from "./EditorService3";