+++ src/editor-service/index.ts
@@ -1,0 +1,2 @@
+export * from "./createEditorService";
+export * from "./EditorService";